requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.8.0
certifi==2019.6.16        # via requests
chardet==3.0.4            # via readability-lxml, requests
cssselect==1.0.3          # via readability-lxml
idna==2.8                 # via requests
lxml==4.4.0               # via readability-lxml
readability-lxml==0.7.1
requests==2.22.0
soupsieve==1.9.2          # via beautifulsoup4
urllib3==1.25.3           # via requests
